Refactoring integration and unit tests
- Separate projects for Integration Tests and Unit Tests
- Use IntegrationTestsSetup class as IClassFixture for AuthenticationControllerTests
