+++ RestaurantSimulation.Backend/RestaurantSimulation.Backend.sln
@@ -13,7 +13,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RestaurantSimulation.Domain
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RestaurantSimulation.Infrastructure", "RestaurantSimulation.Infrastructure\RestaurantSimulation.Infrastructure.csproj", "{FA061141-A7BC-4B98-94FC-D933A7B66C14}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RestaurantSimulation.Tests", "RestaurantSimulation.Tests\RestaurantSimulation.Tests.csproj", "{DE84168A-5C0C-4B55-937C-9B163E5697CC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RestaurantSimulation.IntegrationTests", "RestaurantSimulation.IntegrationTests\RestaurantSimulation.IntegrationTests.csproj", "{68BC843E-BDB0-4733-96BA-D6A5FF370F52}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RestaurantSimulation.UnitTests", "RestaurantSimulation.UnitTests\RestaurantSimulation.UnitTests.csproj", "{3D82E907-903B-4F60-9D16-B3D33A9E363F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -41,10 +43,14 @@ Global
 		{FA061141-A7BC-4B98-94FC-D933A7B66C14}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FA061141-A7BC-4B98-94FC-D933A7B66C14}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FA061141-A7BC-4B98-94FC-D933A7B66C14}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DE84168A-5C0C-4B55-937C-9B163E5697CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DE84168A-5C0C-4B55-937C-9B163E5697CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DE84168A-5C0C-4B55-937C-9B163E5697CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DE84168A-5C0C-4B55-937C-9B163E5697CC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{68BC843E-BDB0-4733-96BA-D6A5FF370F52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{68BC843E-BDB0-4733-96BA-D6A5FF370F52}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{68BC843E-BDB0-4733-96BA-D6A5FF370F52}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{68BC843E-BDB0-4733-96BA-D6A5FF370F52}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3D82E907-903B-4F60-9D16-B3D33A9E363F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3D82E907-903B-4F60-9D16-B3D33A9E363F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3D82E907-903B-4F60-9D16-B3D33A9E363F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3D82E907-903B-4F60-9D16-B3D33A9E363F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
