+++ RestaurantSimulation.Backend/RestaurantSimulation.IntegrationTests/Restaurant/RestaurantMenuCategory/MenuCategoryControllerTests.cs
@@ -1,0 +1,6 @@
+﻿namespace RestaurantSimulation.IntegrationTests.Restaurant.RestaurantMenuCategory
+{
+    public class MenuCategoryControllerTests
+    {
+    }
+}
