+++ RestaurantSimulation.Backend/RestaurantSimulation.IntegrationTests/RestaurantSimulation.IntegrationTests.csproj
@@ -1,0 +1,39 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+	  <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="6.0.8" />
+	  <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="6.0.8" />
+	  <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
+	  <PackageReference Include="Moq" Version="4.18.2" />
+	  <PackageReference Include="RestSharp" Version="108.0.1" />
+	  <PackageReference Include="Shouldly" Version="4.1.0" />
+	  <PackageReference Include="WebMotions.Fake.Authentication.JwtBearer" Version="6.1.0" />
+	  <PackageReference Include="xunit" Version="2.4.1" />
+	  <PackageReference Include="Xunit.Extensions.Ordering" Version="1.4.5" />
+	  <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
+		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		  <PrivateAssets>all</PrivateAssets>
+	  </PackageReference>
+	  <PackageReference Include="coverlet.collector" Version="3.1.2">
+		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		  <PrivateAssets>all</PrivateAssets>
+	  </PackageReference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\RestaurantSimulation.Api\RestaurantSimulation.Api.csproj" />
+    <ProjectReference Include="..\RestaurantSimulation.Application\RestaurantSimulation.Application.csproj" />
+    <ProjectReference Include="..\RestaurantSimulation.Contracts\RestaurantSimulation.Contracts.csproj" />
+    <ProjectReference Include="..\RestaurantSimulation.Domain\RestaurantSimulation.Domain.csproj" />
+    <ProjectReference Include="..\RestaurantSimulation.Infrastructure\RestaurantSimulation.Infrastructure.csproj" />
+  </ItemGroup>
+
+</Project>
