+++ RestaurantSimulation.Backend/RestaurantSimulation.IntegrationTests/Usings.cs
@@ -1,0 +1,8 @@
+global using Xunit;
+
+//Optional
+[assembly: CollectionBehavior(DisableTestParallelization = true)]
+//Optional
+[assembly: TestCaseOrderer("Xunit.Extensions.Ordering.TestCaseOrderer", "Xunit.Extensions.Ordering")]
+//Optional
+[assembly: TestCollectionOrderer("Xunit.Extensions.Ordering.CollectionOrderer", "Xunit.Extensions.Ordering")]
