+++ RestaurantSimulation.Backend/RestaurantSimulation.UnitTests/Authentication/Commands/UpdateUserHandlerTests.cs
@@ -3,11 +3,11 @@ using RestaurantSimulation.Application.Authentication.Commands.UpdateUser;
 using RestaurantSimulation.Application.Authentication.Common.Services.ExtractUserClaims;
 using RestaurantSimulation.Application.Common.Interfaces.Persistence;
 using RestaurantSimulation.Domain.Entities.Authentication;
-using RestaurantSimulation.Tests.UnitTests.Mocks.Authentication;
+using RestaurantSimulation.UnitTests.Mocks.Authentication;
 using Shouldly;
 using RestaurantSimulation.Domain.RestaurantApplicationErrors;
 
-namespace RestaurantSimulation.Tests.UnitTests.Authentication.Commands
+namespace RestaurantSimulation.UnitTests.Authentication.Commands
 {
     public class UpdateUserHandlerTests
     {
