+++ RestaurantSimulation.Backend/RestaurantSimulation.UnitTests/RestaurantSimulation.UnitTests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -17,6 +17,7 @@
     <PackageReference Include="Shouldly" Version="4.1.0" />
     <PackageReference Include="WebMotions.Fake.Authentication.JwtBearer" Version="6.1.0" />
     <PackageReference Include="xunit" Version="2.4.1" />
+    <PackageReference Include="Xunit.Extensions.Ordering" Version="1.4.5" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
@@ -33,7 +34,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="UnitTests\Mocks\Restaurant\" />
+    <Folder Include="Mocks\Restaurant\" />
   </ItemGroup>
 
 </Project>
